Separação dos testes por CST
